Soft merge with branch 'Allan'

+++ backend/src/main/java/com/auu_sw3_6/Himmerland_booking_software/service/ResourceService.java
@@ -95,8 +95,13 @@ public abstract class ResourceService<T extends Resource> {
         return repository.save(updatedResource);
     }
 
-    public void deleteResource(Long id) {
-        repository.deleteById(id);
+    public boolean deleteResource(Long id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public Optional<byte[]> getResourcePictureByResourcename(String resourceName) {
